Delete Channel Groups and add Test

+++ bundles/org.openhab.binding.liquidcheck/src/main/java/org/openhab/binding/liquidcheck/internal/json/Measure.java
@@ -2,7 +2,7 @@ package org.openhab.binding.liquidcheck.internal.json;
 
 public class Measure {
     public double level;
-    public double content;
+    public int content;
     public Raw raw;
     public int age;
 }
